LayoutTests/svg/custom/svg-parse-overflow-3.html

<html>
<body>
<script>
document.body.offsetTop;
if (window.layoutTestController)
    layoutTestController.dumpAsText(true);
</script>
<style>
svg {
stroke-dasharray: 400;
stroke: 10px;}
</style>
<svg>
<rect width="100" height="105000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000">
</svg>
</body>
