LayoutTests/svg/custom/svg-parse-overflow-3-expected.txt

CONSOLE MESSAGE: line 13: Error: Invalid value for <rect> attribute height="105000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
